04_dbt_project/docker-requirements.txt
fastparquet
pandas
prefect-sqlalchemy==0.2.2
prefect-gcp[cloud_storage]==0.2.3
protobuf==4.21.11
dbt-postgres
dbt-bigquery